+++ pyodide-build/pyodide_build/common.py
@@ -12,13 +12,49 @@ def _parse_package_subset(query: Optional[str]) -> Optional[Set[str]]:
     Also add the list of mandatory packages: ["pyparsing", "packaging",
     "micropip"]
 
+    Supports folowing meta-packages,
+     - 'core': corresponds to packages needed to run the core test suite
+       {"micropip", "pyparsing", "pytz", "packaging", "Jinja2"}. This is the default option
+       if query is None.
+     - 'min-scipy-stack': includes the "core" meta-package as well as some of the
+       core packages from the scientific python stack and their dependencies:
+       {"numpy", "scipy", "pandas", "matplotlib", "scikit-learn", "joblib", "pytest"}.
+       This option is non exaustive and is mainly intended to make build faster
+       while testing a diverse set of scientific packages.
+     - '*': corresponds to all packages (returns None)
+
+    Note: None as input is equivalent to PYODIDE_PACKAGES being unset and leads
+    to only the core packages being built.
+
     Returns:
-      a set of package names to build or None.
+      a set of package names to build or None (build all packages).
     """
     if query is None:
-        return None
+        query = "core"
+
+    core_packages = {"micropip", "pyparsing", "pytz", "packaging", "Jinja2"}
+    core_scipy_packages = {
+        "numpy",
+        "scipy",
+        "pandas",
+        "matplotlib",
+        "scikit-learn",
+        "joblib",
+        "pytest",
+    }
     packages = {el.strip() for el in query.split(",")}
     packages.update(["pyparsing", "packaging", "micropip"])
+    # handle meta-packages
+    if "*" in packages:
+        # build all packages
+        return None
+    elif "core" in packages:
+        packages |= core_packages
+        packages.discard("core")
+    elif "min-scipy-stack" in packages:
+        packages |= core_packages | core_scipy_packages
+        packages.discard("min-scipy-stack")
+
     # Hack to deal with the circular dependence between soupsieve and
     # beautifulsoup4
     if "beautifulsoup4" in packages: