+++ docs/project/changelog.md
@@ -23,8 +23,16 @@ substitutions:
 - {{Fix}} The "long output truncated" message now appears on a separate line as intended.
   {pr}`1814`
 
+### Python / JavaScript type conversions
+
 - {{Fix}} Conversion of very large strings from Javascript to Python works
   again. {pr}`1806`
+
+### pyodide-build
+
+- By default only a minimal set of packages is built. To build all packages set
+  `PYODIDE_PACKAGES='*'` In addition, `make minimal` was removed, since it is
+  now equivalent to `make` without extra arguments. {pr}`1801`
 
 - {{Fix}} The `setInterruptBuffer` command is now publically exposed again, as
   it was before.